Break out plan & context, fix tests
